auth/requirements.txt
FLASK==1.1.2
FLASK_SQLALCHEMY==2.4.4
FLASK_LOGIN==0.5.0
WERKZEUG
gunicorn==20.0.4
psycopg2-binary
Flask-SQLAlchemy==2.4.4
PIKA
python-dotenv==0.14.0
click
PYSMB==1.2.6